blog_practice/requirements.txt
asgiref==3.5.2
Django==4.0.5
django-phonenumber-field==6.3.0
sqlparse==0.4.2
tzdata==2022.1
